# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "CancelCancellingResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "const": "cancelling",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id"
        ],
        "title": "CancelCancellingResponse",
        "type": "object"
      },
      "CancelNoActiveGenerationResponse": {
        "properties": {
          "status": {
            "const": "no_active_generation",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "CancelNoActiveGenerationResponse",
        "type": "object"
      },
...
      "GenerateVideoCancelledResponse": {
        "properties": {
          "status": {
            "const": "cancelled",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "GenerateVideoCancelledResponse",
        "type": "object"
      },
      "GenerateVideoCompleteResponse": {
        "properties": {
          "status": {
            "const": "complete",
            "title": "Status",
            "type": "string"
          },
          "video_path": {
            "title": "Video Path",
            "type": "string"
          }
        },
        "required": [
          "status",
          "video_path"
        ],
        "title": "GenerateVideoCompleteResponse",
        "type": "object"
      },
      "GenerateVideoModelsSpecsResponse": {
        "properties": {
          "api_models": {
            "items": {
              "$ref": "#/components/schemas/LTXVideoGenerationModelSpecItem"
            },
            "title": "Api Models",
            "type": "array"
          },
          "local_models": {
            "items": {
              "$ref": "#/components/schemas/LTXVideoGenerationModelSpecItem"
            },
            "title": "Local Models",
            "type": "array"
          }
        },
        "required": [
          "local_models",
          "api_models"
        ],
        "title": "GenerateVideoModelsSpecsResponse",
        "type": "object"
      },
      "GenerateVideoRequest": {
        "properties": {
          "aspectRatio": {
            "default": "16:9",
            "enum": [
              "16:9",
              "9:16"
            ],
            "title": "Aspectratio",
            "type": "string"
          },
          "audio": {
            "default": false,
            "title": "Audio",
            "type": "boolean"
          },
          "audioPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audiopath"
          },
          "cameraMotion": {
            "default": "none",
            "enum": [
              "none",
              "dolly_in",
              "dolly_out",
              "dolly_left",
              "dolly_right",
              "jib_up",
              "jib_down",
              "static",
              "focus_shift"
            ],
            "title": "Cameramotion",
            "type": "string"
          },
          "duration": {
            "anyOf": [
              {
                "enum": [
                  2,
                  3,
                  4,
                  5,
                  6,
                  8,
                  10,
                  12,
                  14,
                  16,
                  18,
                  20
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 5,
            "title": "Duration"
          },
          "fps": {
            "default": 24,
            "enum": [
              24,
              25,
              48,
              50
            ],
            "title": "Fps",
            "type": "integer"
          },
          "imagePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagepath"
          },
          "keyframes": {
            "items": {
              "$ref": "#/components/schemas/KeyframeInput"
            },
            "title": "Keyframes",
            "type": "array"
          },
          "lastImagePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastimagepath"
          },
          "loras": {
            "items": {
              "$ref": "#/components/schemas/LoraEntry"
            },
            "title": "Loras",
            "type": "array"
          },
          "model": {
            "default": "fast",
            "enum": [
              "fast",
              "pro",
              "fast-2.5",
              "pro-2.5"
            ],
            "title": "Model",
            "type": "string"
          },
          "negativePrompt": {
            "default": "",
            "title": "Negativeprompt",
            "type": "string"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "resolution": {
            "default": "1080p",
            "enum": [
              "540p",
              "720p",
              "1080p",
              "1440p",
              "2160p"
            ],
            "title": "Resolution",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "GenerateVideoRequest",
        "type": "object"
      },
      "GenerationProgressResponse": {
        "properties": {
          "cancellable": {
            "title": "Cancellable",
            "type": "boolean"
          },
          "currentStep": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentstep"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "idle",
              "running",
              "complete",
              "cancelled",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "totalSteps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalsteps"
          }
        },
        "required": [
          "status",
          "phase",
          "progress",
          "currentStep",
          "totalSteps",
          "cancellable"
        ],
        "title": "GenerationProgressResponse",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "HuggingFaceAuthStatusResponse": {
        "properties": {
          "status": {
            "enum": [
              "authenticated",
              "pending",
              "not_authenticated"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HuggingFaceAuthStatusResponse",
        "type": "object"
      },
      "HuggingFaceLoginResponse": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "code_challenge": {
            "title": "Code Challenge",
            "type": "string"
          },
          "code_challenge_method": {
            "title": "Code Challenge Method",
            "type": "string"
          },
          "redirect_uri": {
            "title": "Redirect Uri",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "redirect_uri",
          "scope",
          "state",
          "code_challenge",
          "code_challenge_method"
        ],
        "title": "HuggingFaceLoginResponse",
        "type": "object"
      },
      "HuggingFaceLogoutResponse": {
        "properties": {
          "status": {
            "const": "logged_out",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HuggingFaceLogoutResponse",
        "type": "object"
      },
...
      "IcLoraExtractRequest": {
        "properties": {
          "conditioning_type": {
            "default": "canny",
            "enum": [
              "canny",
              "depth"
            ],
            "title": "Conditioning Type",
            "type": "string"
          },
          "frame_time": {
            "default": 0,
            "title": "Frame Time",
            "type": "number"
          },
          "video_path": {
            "title": "Video Path",
            "type": "string"
          }
        },
        "required": [
          "video_path"
        ],
        "title": "IcLoraExtractRequest",
        "type": "object"
      },
      "IcLoraExtractResponse": {
        "properties": {
          "conditioning": {
            "title": "Conditioning",
            "type": "string"
          },
          "conditioning_type": {
            "enum": [
              "canny",
              "depth"
            ],
            "title": "Conditioning Type",
            "type": "string"
          },
          "frame_time": {
            "title": "Frame Time",
            "type": "number"
          },
          "original": {
            "title": "Original",
            "type": "string"
          }
        },
        "required": [
          "conditioning",
          "original",
          "conditioning_type",
          "frame_time"
        ],
        "title": "IcLoraExtractResponse",
        "type": "object"
      },
      "IcLoraGenerateCancelledResponse": {
        "properties": {
          "status": {
            "const": "cancelled",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "IcLoraGenerateCancelledResponse",
        "type": "object"
      },
      "IcLoraGenerateCompleteResponse": {
        "properties": {
          "status": {
            "const": "complete",
            "title": "Status",
            "type": "string"
          },
          "video_path": {
            "title": "Video Path",
            "type": "string"
          }
        },
        "required": [
          "status",
          "video_path"
        ],
        "title": "IcLoraGenerateCompleteResponse",
        "type": "object"
      },
...
      "KeyframeInput": {
        "description": "CamelCase of LTXV keyframe-edit `{image_uri, frame_index, strength}`.",
        "properties": {
          "frameIndex": {
            "minimum": 0.0,
            "title": "Frameindex",
            "type": "integer"
          },
          "imagePath": {
            "title": "Imagepath",
            "type": "string"
          },
          "strength": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Strength",
            "type": "number"
          }
        },
        "required": [
          "imagePath",
          "frameIndex"
        ],
        "title": "KeyframeInput",
        "type": "object"
      },
...
      "LTXVideoGenerationModelSpecItem": {
        "properties": {
          "pipeline": {
            "enum": [
              "fast",
              "pro",
              "fast-2.5",
              "pro-2.5"
            ],
            "title": "Pipeline",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/LTXVideoGenerationSpec"
          }
        },
        "required": [
          "pipeline",
          "spec"
        ],
        "title": "LTXVideoGenerationModelSpecItem",
        "type": "object"
      },
      "LTXVideoGenerationResolutionSpec": {
        "properties": {
          "fps_to_durations": {
            "additionalProperties": {
              "items": {
                "enum": [
                  2,
                  3,
                  4,
                  5,
                  6,
                  8,
                  10,
                  12,
                  14,
                  16,
                  18,
                  20
                ],
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Fps To Durations",
            "type": "object"
          }
        },
        "required": [
          "fps_to_durations"
        ],
        "title": "LTXVideoGenerationResolutionSpec",
        "type": "object"
      },
      "LTXVideoGenerationSpec": {
        "properties": {
          "a2v_supported_resolutions_durations": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/LTXVideoGenerationResolutionSpec"
                },
                "propertyNames": {
                  "enum": [
                    "540p",
                    "720p",
                    "1080p",
                    "1440p",
                    "2160p"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "A2V Supported Resolutions Durations"
          },
          "capabilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LTXOfferingCapabilitiesSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "supported_resolutions_durations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LTXVideoGenerationResolutionSpec"
            },
            "propertyNames": {
              "enum": [
                "540p",
                "720p",
                "1080p",
                "1440p",
                "2160p"
              ]
            },
            "title": "Supported Resolutions Durations",
            "type": "object"
          }
        },
        "required": [
          "display_name",
          "supported_resolutions_durations"
        ],
        "title": "LTXVideoGenerationSpec",
        "type": "object"
      },
...
      "LoraEntry": {
        "properties": {
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "scale": {
            "default": 1.0,
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Scale",
            "type": "number"
          }
        },
        "required": [
          "ref"
        ],
        "title": "LoraEntry",
        "type": "object"
      },
...
      "LtxInsufficientFundsErrorResponse": {
        "properties": {
          "code": {
            "const": "LTX_INSUFFICIENT_FUNDS",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "LtxInsufficientFundsErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LTX-2 Video Generation Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/huggingface/callback": {
      "get": {
        "operationId": "route_hf_callback_api_auth_huggingface_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "default": "",
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "default": "",
              "title": "Error",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Hf Callback",
        "tags": [
          "hf_auth"
        ]
      }
    },
    "/api/auth/huggingface/login": {
      "post": {
        "operationId": "route_hf_login_api_auth_huggingface_login_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HuggingFaceLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Hf Login",
        "tags": [
          "hf_auth"
        ]
      }
    },
    "/api/auth/huggingface/logout": {
      "post": {
        "operationId": "route_hf_logout_api_auth_huggingface_logout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HuggingFaceLogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Hf Logout",
        "tags": [
          "hf_auth"
        ]
      }
    },
    "/api/auth/huggingface/status": {
      "get": {
        "operationId": "route_hf_auth_status_api_auth_huggingface_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HuggingFaceAuthStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Hf Auth Status",
        "tags": [
          "hf_auth"
        ]
      }
    },
...
    "/api/generate": {
      "post": {
        "description": "POST /api/generate — video generation from JSON body.",
        "operationId": "route_generate_api_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GenerateVideoCompleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GenerateVideoCancelledResponse"
                    }
                  ],
                  "title": "Response Route Generate Api Generate Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LtxInsufficientFundsErrorResponse"
                }
              }
            },
            "description": "LTX API credits are insufficient for the requested generation"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Generate",
        "tags": [
          "generation"
        ]
      }
    },
...
    "/api/generate/cancel": {
      "post": {
        "description": "POST /api/generate/cancel.",
        "operationId": "route_generate_cancel_api_generate_cancel_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CancelCancellingResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CancelNoActiveGenerationResponse"
                    }
                  ],
                  "title": "Response Route Generate Cancel Api Generate Cancel Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Generate Cancel",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generate/models-specs": {
      "get": {
        "description": "GET /api/generate/models-specs.",
        "operationId": "route_generate_model_specs_api_generate_models_specs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateVideoModelsSpecsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Generate Model Specs",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/progress": {
      "get": {
        "description": "GET /api/generation/progress.",
        "operationId": "route_generation_progress_api_generation_progress_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Generation Progress",
        "tags": [
          "generation"
        ]
      }
    },
...
    "/api/ic-lora/extract-conditioning": {
      "post": {
        "operationId": "route_ic_lora_extract_api_ic_lora_extract_conditioning_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IcLoraExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcLoraExtractResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Ic Lora Extract",
        "tags": [
          "ic-lora"
        ]
      }
    },
    "/api/ic-lora/generate": {
      "post": {
        "operationId": "route_ic_lora_generate_api_ic_lora_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IcLoraGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IcLoraGenerateCompleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/IcLoraGenerateCancelledResponse"
                    }
                  ],
                  "title": "Response Route Ic Lora Generate Api Ic Lora Generate Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Ic Lora Generate",
        "tags": [
          "ic-lora"
        ]
      }
    },
...
  }
}